init: Get the kernel stack up and running
locore.s defines proc0paddr, which is the p_addr member of struct
proc and is of type (struct user *). It is used to handle MD state
of a process. It is also abused as kernel stack (check comments at
the member in proc.h). While the user struct grows bottom up, the
kernel stack grows top-down. The stack size is USPACE large (which
resolves to 2*PAGE_SIZE), thus we have 8kB kernel stack.

NOTE: On MULTIPROCESSOR machines there is a kernel stack for each
      CPU, but we do not bother for now.