mem: Reservce space for (A)PTD
OpenBSD originally maintains its (alternate) pagetable directory
(A)PTD in memory at virtual address 0xcff3f000 (and 0xfffff000).

Since we are running as a simple programm, we do not have the
opportunity to simply write to that location (Fiasco is loaded there
anyway). Simply making enough room in the .data section should resolve
that problem for now.

Of course, we still need to adjust all macros in pmap.h.