mem: Relocate main memory to 0x40000000 and add l4lxlib memory manager
Relocating the main memory to 0x40000000 has the benefit of easing the
physical to virtual kernel address translation. As we have moved the kernel
to 0x10000000, we now can compute

 vaddr_t = paddr_ - 0x10000000

Additionally, I implemented a first step in emulating the locore.s behaviour
in populating an initial page directory and initial page maps.