l4ser: Get capability directly from Fiasco instead of MOE
Get an input/output capability directly from the Fiasco kernel. We used to
have one from MOE before, but MOE is not able to handle input correctly.

WARNING: This is a hack and shows a design flaw in the L4re system. The
politically correct way of doing things would be to have a tty mutliplexer
service running on Fiasco, which we may ask to provide the I/O service for
us.

Input is still untested.