bios: Fake default PC BIOS initialization
Fix the bios(4) device initialization by pretending to be a default PC-based
BIOS. We claim to not have anything to initialize.

Since there is no easy way to build a kernel without a bios(4) device, we
have two options here.

 1. Let the device probe fail.
 2. Pretend a device.

We do number 2. There are some hard-coded references to code in this driver
and I do not know if they need a functioning device in the device tree. If
everything else fails, implementing option 1 is rather easy.