segments: Add a dummy global descriptor table for the boot CPU
This puts some new allocated memory as GDT. The original implementation just
wants to have some VA space and maps the bootstrap GDT itself, which we cannot
do. The result is the same, as long as nobody ever touches the bootstrap GDT
directly, but uses its pointer (which seems to be the case, at a grep).