cpu: Disable privileged instructions
We disable a bunch of privileged instructions, which cannot be executed
as a normal process. The rcr?() class of functions always return 0. This
is not a problem, since the most common use case is

  lcr4(rcr4() | CR4_PAE);

The only side effect is that we do not have a complete register set in our
user struct, but I think we can live with that fact.