clock: Implement delay(9) on L4
We use an RPC timeout to do the busy-waiting. This should save some CPU cycles
on the host. I am not even sure, if we really need delay(9) here at all, since
it is mostly used in device drivers to let the hardware settle. But just in case,
we have it.