clock: Implement stub initclocks() on L4
This is a stub for initclocks() rehosted to L4.

We implement a timer device already. We need this "device", since we are messing
around inside the i386 architecture. That makes it impossible to

  #undef __HAVE_TIMECOUNTER

Therefore we need to implement that device. Otherwise implementing microtime(9)
would have been sufficient.

Later on (see TODO) we should setup IRQ handling on the vCPU and handle these IRQs.