irq: Attach IRQ handler registration on vCPU and enable timer interrupt
Attach IRQ handler registration for L4 to isa_intr_establish(). Register
the timer IRQ (IRQ0) handler and run an L4 thread to interrupt the vCPU-
enabled main thread hz times per second.

We respect splx(9). Interrupts firing on higher SPLs notify the system and
handlers get executed, when the SPL is lowered again. This of course only
happens if interrupts on the vCPU are enabled.