irq: Prepare trapframe when re-entering deferred interrupts for hardclock(9)
We use the current proc's p_md.md_regs trapframe registers on interrupt entry
to save the vCPU state. Now, we actually save a correct state for tf_cs, so
that evaluating the mode of operation does not need too much adjustment. Now
we can elaborate, if we are running in kernelmode or usermode. vCPU saving
operations are basically the same, as is in L4/Linux. The only thing here is,
that we have proper names for the constants.

When re-entering a deferred interrupt at e.g. splx(), we need to prepare that
trapframe again, instead of just using a fake one. I assume we _always_ have
a valid curproc. Statistical information will thus be charged for curproc,
every time.

Since we use a struct trapframe for every vCPU entry, hardclock(9) needs to
use a struct trapframe, too. OpenBSD/i386 uses a struct intrframe for that.