vcpu: Enable vCPU mode earlier
Since there are many functions out there, which rely on proper spl(9) behaviour,
turn on vCPU mode earlier in the boot process. This way, we make sure, that faults
are handled by our own vCPU handler function, instead of erroring out.

That change is a preparation for loading the initial ramdisk.