rd: Build and load an initial ramdisk
Allow an initial ramdisk, provided by the lx_rd command line parameter to the
kernel image, to be loaded on boot. Grub should load the ramdisk image, so that
we can access the according dataspace. We copy the image to the location, where
OpenBSD searches it.

The kernel option RAMDISK_HOOKS enables loading an initial ramdisk. The kernel
option MINIROOTSIZE tells us the space needed for the initial ramdisk.

The initial ramdisk is built from this same source tree. A mini howto is provided
in the base directory at src/README.L4.