irq: Do not enter the Fiasco kernel with IRQs enabled on the vCPU
Restructure the IRQ handling, since we are not supposed to enter the
Fiasco kernel with IRQs enabled.

Therefore we need to wrap all Fiasco system calls with the restructured
L4XV_? macros. These save the current vCPU IRQ state and restore it later.
l4lx_* functions need to be protected alltogether. For debugging purposes
a new l4x_printf() -- wrapping LOG_printf() is introduced.

Also, the startup sequence was restructured. We leave IRQs off now, as
long as we can. One-time initialization functions are supposed to run
before enabling IRQs.