trap: Add basic page fault handling
OpenBSD/i386 uses the trap() function for handling hardware exceptions, as well
as page faults. Pour in the vCPU state into that function.

For page fault handling, obvioulsy we need the page fault address (pfa). Hook up
the default crc2() to provide us with the current vCPU pfa, so we do not need to
munge on the general trap handler all that much.