debug: Make the OpenBSD kernel debugger DDB invoke correctly
The OpenBSD/i386 kernel debugger normally runs on an INT3. Since we are not running
on bare metal, Fiasco's kernel debugger also runs on INT3. So every voluntary
invocation of the kernel debugger started the Fiasco kernel debugger.

This is bad for some reasons. Fiasco does not know of any of OpenBSD's kernel internal
data structures. This is bad for debugging, especially when some generic OpenBSD
code wants us to se what we have screwed up on OpenBSD's side.

As a later step, we might want to add a module to DDB/KGDB to invoke the Fiasco
kernel debugger.

I am not quite sure about the side effects of not going through an interrupt
cycle here. We might loose some information.