copy: Implement substitutes for the copy(9) family of functions
The copy(9) family of functions are the interface between kernel and userspace.
They are used for data exchange. Since we cannot access userspce memory directly,
we walk the current process's (curproc) page directory (pmap->pm_pdir), to find
the according pages in userspace. We can directly copy our data there, keeping in
mind that each page is situated in a different place in our physical memory area.

If the pmap does not (yet) know about the page mapping in userspace, we run the
page fault routine uvm_fault() to resolve it and (hopefully) enter the correct
mapping for us. After that, we can translate the userspace virtual address into
the physical address we own in our address space and copy the data there on a per
page basis.

In a later step, L4 will map the physical page into the process's address space,
according to the pmap mapping.