irq: Acknowledge pending interrupts directly after handling them
Previously we ran the IRQ handler routines multiple times on certain occasions.
With this path, we acknowledge interrupts directly after having run the according
interrupt service routines. In this way it does not matter how we enter the IRQ
handling routines, meaning we do not care if we handle real interrupts, or spl(9).

Also, fix it while here.