copy: Do not use otherwise unused registers in assembler
Initially, I chose to use register %ebx, because it was otherwise unused in the
copy(9) family of string functions. That was a bad idea, because it showed cruel
behaviour. I guess GCC does not catch that the register was used and uses our new
assigned value further on.

Nevertheless, I got strange errors that code following these copy(9) functions
behaved strangely.

Using a register, which is used in OpenBSD/i386 as well makes these bugs disappear.