fork: Copy trapframe of old process to new process
Copy all userland register contents from the old process to the new one.
Therefore, if the new process is scheduled to run and exits to userland,
it will start running right at the same position as the old one.

One exception is that the new process executes child_return() on the
interrupt exit path, resetting %eax.