l4ser: Rewrite the serial driver for L4
Up until now, we only had the consdev working. This implements the
cdev l4ser as a tty device.

Normally OpenBSD requires tty devices to register an interrupt handler
routine. We avoid that on the read path by injecting the character
from the L4 IPC call into the tty line. On the write path, the generic
tty handler function l_write disassembles the uio for us and we get
triggered from it as l4serstart() where we actually put it on the line.