clock: Change interrupt type from PULSE to EDGE on L4
Since IST_PULSE is incompatible with any other type of interrupt, we change
the type of the clock interrupt here. On L4 this does not even matter at all.

This commit is in preparation to the serial driver, which needs to hook on
the clock interrupt to look for keypresses.