l4ser: Implement a full-featured serial driver for L4
This patch implements a full-featured serial driver for L4. We use the already
implemented console functions as they already fit our needs.

The output path looks like the following. The driver layer calls l4serwrite()
for new output. This method converts the uio input into a TTY-compatible stream
of chars. On open(2), we registered l4serstart() as *t_oproc to start the output.
The TTY layer will call *t_oproc after l4serwrite(). In this function, we use
the console function to output one charcter of the character stream at a time.
After that we try to wake up waiting processes.

The input path is realized through an interrupt service routine. We hook on the
clock interrupt (IRQ0) and poll for new keys from L4 on every interrupt. For now,
we only fetch 20 characters at a time, so that the kernel does not get hogged.
After we finish fetching keys, we try to wake up waiting processes again.