vCPU: Only ever maipulate vCPU save state area with IRQs disabled
The vCPU save state area is not re-entrant safe. Therefore we only must manipulate
that state in either of the two following cases, with IRQ event delivery disabled:

1. On vCPU entry

On vCPU entry we need to get the current vCPU save state to copy the trapframe.
OTOH we need to see, if we came from userland or kernel land.

2. On vCPU exit

When returning to the interrupted context, we need to adjust the save state area.

In either case, the current state shall only ever be revealed through the current
trapframe. With interrupts enabled, informartion from the vCPU save state area is
useless!