irq: Fake kernel mode for in-kernel interrupt handling (again)
Stupid me broke that in the last commit. Note that do_vcpu_irq() only ever gets
called in kernel context as event handler for l4vcpu library calls. Therefore it
is safe to fake a kernel code segment in curproc's trapframe. We will never resume
execution in userland with it. In theory, we would need to raise a vCPU event on
our own here.