vcpu: Handle page faults from the very beginning
When enabling vCPU mode, we did not handle page faults. Instead our parent thread
tried to handle them for us. Under normal circumstances this is not a problem and
pagefaults should not happen that early in the boot process anyway. But not enabling
it makes debugging a lot harder. The vCPU saves the register state for us. On the
other hand, enabling page fault handling earlier does not cost anything.