mem: Use superpages for main memory allocation if applicable
If, by any means, the effective main memory size fits into superpages, use these
for allocation. This is rarely the case, but once it is, it lowers TLB thrashing.

Maybe we should not allocate the effective main memory, but the requested memory
instead, so that it is clear from the kernel configuration, if superpages can be
used.