proc: Set process name for Fiasco kernel debugger
Normally, we would only need to set the process name twice. On fork(2), we can
set the parent's process name, since both run in the same address space. Later
on exec(2), we would need to set the final name of that process. Unfortunately
the name is copied into the proc structure after the latest MD code got executed.
Therefore, we set the name on every process switch.