rd: Do not copy ramdisk contents into the kernel image
Since we load a separate ramdisk image anyway, there is no need to copy it into
the kernel's rd driver. We just alter the pointers in a way that the rd driver is
happy now. Especially when dealing with 512MB ramdisk images this will be no fun.

NOTE: We are changing MI code here. This was not supposed to happen, but it makes
      all our lives easier.