trap: Rework usermode pagefault handling
The former design was flawed. We need to completely re-evaluate the page tables,
when the page fault handling function uvm_fault() kicked in. That means, we have
to restart the walk.

While here, rename the function to properly reflect, that we walk the page tables.
This functionality is cheap, if OpenBSD already entered the pages (e.g. first copying
out data on a page, then mapping it in a task later).

Also, fix the vCPU resume path. Always pass 0 as fp1, if we do not intend to send
a page to the returning process.