L4: Handle L4-specific page faults
Sigma0 provides us with memory at boot time, but it does not guarantee that it
is mapped all the time. In theory, memory can vanish without any reason. This is
especially true for the memory used as physical memory for programs.

We just need to touch the faulting page and continue execution.