l4ser: Make line reading work for echo terminal types
If a terminal is set to "echo" with

 # stty echo

all subsequent calls to

 # read a

should print the read characters on the tty. The current implementation missed
to wake up the terminal writers and print the characters. With terminals in the
"-echo" state, the patch also works and does not print anything. This is because
there is no process in the writer queue.