pmap: Only unmap KVAs when there was a mapping before
When removing a KVA mapping, plain OpenBSD/i386 does not need to care if that mapping
already exists. It just removes the pte mapping by writing a 0 at a memory location
(page tables). This does not work in the same way on Fiasco/L4. We could stop handling
the error case in l4re_rm_detach(). On the other hand, we do save an IPC call when we
do not call l4re_rm_detach() if it is not necessary.