pmap: Stop propagating pmap changes to the L4 task on process exit
When a process exits in OpenBSD, the whole address space gets cleared. Since we destroy
the according L4 task anyway, it is pointless to propagate the address space cleanups
to that task. With this path, all address space changes are not forwarded to the L4
task anymore. This gives a speed gain of factor 3 in lmbench's fork+exit test.