pmap: Fix size of flexpage for l4_task_unmap
When unmapping a page in a different task, the size of the flexpage is given in
log2(size-in-bytes). Therefore, destroying a mapping with page size has a flexpage
size of PGSHIFT. That also means, we cannot easily unmap more pages at once.

Now that we got the size right, Fiasco does not unmap tons of pages for a single
unmap operation anymore. That led to a bunch of problems. The unmap operation took
quite some time. Additionally, we had to page in all the pages again. Both things
were odd.