tc: Use the TSC for a more accurate timing if possible
If the running CPU features a TSC, use it to get a more accurate timing.
Fiasco allows userspace programs to use the TSC.

Fallback to the clock on the kernel information page (KIP), if no TSC
is present and let Fiasco figure out how to provide it. (This is the old
behaviour.)