Licensing: Clarify licensing of all L4 components
Most of the L4 code was derived from L4Linux SVN rev.15. Therefore it is
licensed under the GPLv2, as it was derived from the Linux kernel itself.

All components (e.g. drivers), which are not derived from L4Linux shall
be licensed under the preferred OpenBSD license. At the moment this is ISC.